+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -1107,7 +1107,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M37EHKZRNCGMDQW8N90K7QBB"
+        "tool_01M3BK2DBDEZ05W9PFHW2R5PH5"
     )
 
     display_name: Optional[str] = None
@@ -1412,7 +1412,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M37EHKZQFEXNFXMJHA05T7R7"
+        "tool_01M3BK2DBC668K15WCASFXZW5J"
     )
 
     display_name: Optional[str] = None
@@ -1606,7 +1606,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M37EHKZP1WS7GR1BHTZMEVY3"
+        "tool_01M3BK2DBBAEGRXDP7NJ3KFAPS"
     )
 
     display_name: Optional[str] = None
@@ -1804,7 +1804,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M37EHKZPT0W17F3C2DXST4V6"
+        "tool_01M3BK2DBBB36M562TY3JGNY1Q"
     )
 
     display_name: Optional[str] = None